feat: UI와 로직 분리하여 Header 컴포넌트 리팩토링
