+++ package.json
@@ -10,11 +10,9 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@emoji-mart/data": "^1.2.1",
-    "@emoji-mart/react": "^1.1.1",
     "@tinymce/tinymce-react": "^5.1.1",
     "axios": "^1.7.4",
-    "emoji-mart": "^5.6.0",
+    "emoji-picker-react": "^4.11.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-router-dom": "^6.26.1",
