+++ src/api/recipients.js
@@ -35,7 +35,6 @@ const addMessageToRecipient = async (recipientId, messageData) => {
 // 롤링 페이퍼 대상의 메세지 목록 가져오기
 const getMessagesByRecipientId = async (recipientId) => {
   const response = await axios.get(`/${TEAM}/recipients/${recipientId}/messages/`);
-  console.log('response', response.data);
 
   return response.data;
 };
