+++ src/styles/DefaultLayout.jsx
@@ -1,12 +1,13 @@
 import styled from 'styled-components';
 import { Outlet } from 'react-router-dom';
-import Header from '../components/Header/Header.jsx';
+import HeaderContainer from '../containers/HeaderContainer';
 
 const Layout = styled.div`
   max-width: 1200px;
   margin: 0 auto;
   padding: 20px;
 `;
+
 const MainLayout = styled.div`
   margin-top: 65px;
 `;
@@ -14,7 +15,7 @@ const MainLayout = styled.div`
 export const DefaultLayout = () => {
   return (
     <Layout>
-      <Header />
+      <HeaderContainer />
       <MainLayout>
         <Outlet />
       </MainLayout>
